Improve keychange notice reliability
Bind a single listener to keychange events from the storage interface,
which then looks up relevant conversations and adds notices to them.

Previously we would need to instantiate a conversation model in order to
start listening to its key change events. In practice this usually
happens at startup but we shouldn't rely on it.

// FREEBIE